response-gen/requirements.txt
torch==2.3.0
sentencepiece==0.2.0
transformers[torch]==4.40.0
python-dotenv==1.0.1
fastapi[standard]==0.112.0